Split preparation pipeline
