+++ ssr/input_adapters/adapter_DFC2019.py
@@ -11,9 +11,8 @@ import distutils.dir_util
 
 
 class InputAdapter:
-    def __init__(self, pm: PathManager, preparation_pipeline=None):
+    def __init__(self, pm: PathManager):
         self.pm = pm
-        assert preparation_pipeline is None
 
     def run(self):
         logger.info("Importing the DFC2019 dataset")
