+++ ssr/input_adapters/adapter_MVS3DM.py
@@ -1,19 +1,20 @@
 from ssr.config.ssr_config import SSRConfig
+from ssr.input_adapters.image_extraction_pipeline import ImageExtractionPipeline
 from ssr.path_manager import PathManager
 from ssr.utility.logging_extension import logger
 
 
 class InputAdapter:
-    def __init__(self, pm: PathManager, preparation_pipeline=None):
+    def __init__(self, pm: PathManager):
         self.pm = pm
-        self.preparation_pipeline = preparation_pipeline
         self.config = SSRConfig.get_instance()
 
     def run(self):
         logger.info("Importing the MVS3DM dataset")
 
-        if self.config.extract_msi_pan_image_pairs and self.preparation_pipeline is not None:
-            self.preparation_pipeline.extract_msi_pan_image_pairs(
+        if self.config.extract_msi_pan_image_pairs:
+            image_extraction_pipeline = ImageExtractionPipeline(self.pm)
+            image_extraction_pipeline.extract_msi_pan_image_pairs(
                 extract_msi_pan_image_pairs=self.config.extract_msi_pan_image_pairs,
                 use_consistent_msi_pan_extraction=self.config.use_consistent_msi_pan_extraction,
             )
