+++ ssr/surface_rec/preparation/preparation_pipeline.py
@@ -1,17 +1,9 @@
-import os
 from distutils.dir_util import copy_tree
 from distutils.file_util import copy_file
 from ssr.utility.os_extension import mkdir_safely
 from ssr.utility.os_extension import makedirs_safely
-from ssr.utility.os_extension import assert_dirs_equal
 
 
-from ssr.surface_rec.preparation.data_extraction.extraction_pipeline import (
-    ExtractionPipeline,
-)
-from ssr.config.vissat_config import (
-    create_vissat_extraction_config,
-)
 from ssr.config.ssr_config import SSRConfig
 from ssr.surface_rec.preparation.depth_map_recovery.depth_map_recovery import (
     recover_depth_maps,
@@ -25,103 +17,6 @@ class PreparationPipeline:
     def __init__(self, pm):
         self.pm = pm
         self.ssr_config = SSRConfig.get_instance()
-
-    @staticmethod
-    def extract_files(
-        pan_or_msi_config_fp,
-        ift,
-        oft,
-        execute_parallel,
-        remove_aux_file,
-        apply_tone_mapping,
-        joint_tone_mapping,
-        geo_crop_coordinates=None
-    ):
-
-        pipeline = ExtractionPipeline(pan_or_msi_config_fp)
-        extracted_crops = pipeline.run(
-            ift,
-            oft,
-            execute_parallel,
-            remove_aux_file,
-            apply_tone_mapping,
-            joint_tone_mapping,
-            geo_crop_coordinates=geo_crop_coordinates
-        )
-        return extracted_crops
-
-    def extract_msi_pan_image_pairs(
-        self,
-        extract_msi_pan_image_pairs=True,
-        use_consistent_msi_pan_extraction=True
-    ):
-        if not extract_msi_pan_image_pairs:
-            return
-        # === Additional options for extract_pan and extract_msi === #
-        oft = "png"
-        remove_aux_file = False
-        apply_tone_mapping = True
-        joint_tone_mapping = (
-            False  # Separate tone mapping yields much better results
-        )
-        execute_parallel = True
-
-        pm = self.pm
-        mkdir_safely(pm.ssr_workspace_dp)
-
-        msi_geo_crops = self.extract_msi(pm, oft, execute_parallel, remove_aux_file,
-                                         apply_tone_mapping, joint_tone_mapping)
-
-        if use_consistent_msi_pan_extraction:
-            self.extract_pan(pm, oft, execute_parallel, remove_aux_file,
-                             apply_tone_mapping, joint_tone_mapping, msi_geo_crops)
-        else:
-            self.extract_pan(pm, oft, execute_parallel, remove_aux_file,
-                             apply_tone_mapping, joint_tone_mapping)
-
-    def extract_msi(self, pm, oft, execute_parallel, remove_aux_file, apply_tone_mapping, joint_tone_mapping):
-        mkdir_safely(pm.msi_workspace_dp)
-        create_vissat_extraction_config(
-            pm.msi_config_fp,
-            pm.msi_ntf_idp,
-            pm.msi_workspace_dp,
-            self.ssr_config,
-        )
-        assert os.path.isfile(pm.msi_config_fp)
-
-        msi_geo_crops = PreparationPipeline.extract_files(
-            pm.msi_config_fp,
-            ift="MSI",
-            oft=oft,
-            execute_parallel=execute_parallel,
-            remove_aux_file=remove_aux_file,
-            apply_tone_mapping=apply_tone_mapping,
-            joint_tone_mapping=joint_tone_mapping,
-        )
-        return msi_geo_crops
-
-    def extract_pan(self, pm, oft, execute_parallel, remove_aux_file, apply_tone_mapping, joint_tone_mapping,
-                    msi_geo_crops=None):
-        mkdir_safely(pm.pan_workspace_dp)
-        create_vissat_extraction_config(
-            vissat_config_ofp=pm.pan_config_fp,
-            dataset_dp=pm.pan_ntf_idp,
-            workspace_dp=pm.pan_workspace_dp,
-            ssr_config=self.ssr_config,
-        )
-        assert os.path.isfile(pm.pan_config_fp)
-
-        pan_geo_crops = PreparationPipeline.extract_files(
-            pm.pan_config_fp,
-            ift="PAN",
-            oft=oft,
-            execute_parallel=execute_parallel,
-            remove_aux_file=remove_aux_file,
-            apply_tone_mapping=apply_tone_mapping,
-            joint_tone_mapping=joint_tone_mapping,
-            geo_crop_coordinates=msi_geo_crops
-        )
-        return pan_geo_crops
 
     def run(
         self,
